requirements.txt
streamlit
pillow
python-multipart
